requirements.txt
# requirements.txt
Flask
tensorflow
opencv-python
numpy
google-generativeai
python-dotenv